refactor: 폴더명 변경
-change util to utils
